% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMat.R
\name{DCMat}
\alias{DCMat}
\title{Differential Co-expression}
\usage{
DCMat(exprMat, condition, corMethod = c("spearman", "pearson"))
}
\arguments{
\item{exprMat}{an expression matrix. Rows represent genes and
columns represent samples.}

\item{condition}{a vector that specifies the condition of each sample.
The vector should contain only two conditions. The order of the vector
should match the column order of the expression matrix.}

\item{corMethod}{a character string indicating the correlation coefficient
to use. Can be either "spearman" (default) or "pearson".}
}
\value{
a matrix of z-scores representing the global differential
co-expression network between the two input conditions.
}
\description{
Calculate global differential co-expression network.
}
\details{
This function calculates a global differential co-expression network between
two conditions using a z-score-based method. This method calculates z-scores
through the application of Fisher transformation to the correlation
coefficients and a subsequent z-test.
}
\examples{
# obtain expression data
data(exprBC)
# obtain sample information
data(samplesBC)

# calculate differential correlation matrix
dcMat <- DCMat(exprMat = exprBC,
               condition = samplesBC$er_status,
               corMethod = "spearman")

}
\seealso{
\code{\link{DCPlot}}

\code{\link{eigenCentrality}}
}
\author{
Lanying Wei
}
