% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCPlot.R
\name{DCPlot}
\alias{DCPlot}
\title{Differential Co-expression Network Plotting}
\usage{
DCPlot(
  exprMat,
  condition,
  genes,
  dcMat,
  corMethod = c("spearman", "pearson"),
  posDCCut = 3,
  negDCCut = -posDCCut,
  minimum = 0.2,
  maximum = 1,
  cut = 0.8,
  posCol = c("#4361ee", "#3f37c9"),
  negCol = c("#ff5a5f", "#d90429"),
  layout = c("spring", "circle"),
  repulsion = 1,
  labelCex = 1,
  labelCol = "#6c757d",
  geneList1 = NULL,
  geneList2 = NULL,
  labelCol1 = "#e63946",
  labelCol2 = "#072ac8",
  cond1Title = NULL,
  cond2Title = NULL,
  ...
)
}
\arguments{
\item{exprMat}{an expression matrix. Rows represent genes and
columns represent samples.}

\item{condition}{a vector that specifies the condition of each sample.
The vector should contain only two conditions. The order of the vector
should match the column order of the expression matrix.}

\item{genes}{a character vector of genes to plot.}

\item{dcMat}{a matrix of z-scores representing differential associations
of gene expression between two conditions. See {\link{DCMat}} for more
details. This matrix, along with posDCCut and negDCCut, are used to remove
edges between gene pairs that show low level of differential co-expression.}

\item{corMethod}{a character string indicating the correlation coefficient
to use. Can be either "spearman" (default) or "pearson".}

\item{posDCCut}{a positive numeric value that serves as positive
threshold for DCMat to hide unimportant edges. Gene pairs with DCMat values
below posDCCut and above negDCCut are hidden.}

\item{negDCCut}{a negative numeric value that serves as negative
threshold for DCMat to hide unimportant edges. Gene pairs with DCMat values
below posDCCut and above negDCCut are hidden.}

\item{minimum}{a numeric value between 0 and 1. Edges with absolute
weights less than this value are not shown.}

\item{maximum}{a numeric value between 0 and 1. The highest weight to scale
the edge widths to.}

\item{cut}{a numeric value between 0 and 1. Cut the scaling of edges in
width and color saturation. Edges with absolute weights over this value
have the strongest color intensity and become wider the stronger they are,
and edges with absolute weights under this value have the smallest width
and become vaguer the weaker the weight. If this is set to 0, no cutoff is
used and all edges vary in width and color.}

\item{posCol}{color of positive edges. Can be a vector of two to indicate
color of edges under 'cut' value and color of edges over 'cut' value.}

\item{negCol}{color of negative edges. Can be a vector of two to indicate
color of edges under 'cut' value and color of edges over 'cut' value.}

\item{layout}{the layout of the graph, can be "spring" or "circle". "spring"
gives a force embedded layout; "circle" places all nodes in a single
circle.}

\item{repulsion}{Scalar on the default repulse radius in the spring layout.
Defaults to 1. Setting this argument to lower values (e.g., 0.5) will cause
nodes in the spring layout to repulse each other less. This is especially
useful if a few unconnected nodes cause the giant component to visually
be clustered too much in the same place.}

\item{labelCex}{numeric scalar on the label size.}

\item{labelCol}{Character containing the color of the labels.}

\item{geneList1}{a character vector of a list of genes to be colored.
Genes in this vector are labelled with color 'labelCol1'.}

\item{geneList2}{a character vector of a list of genes to be colored.
Genes in this vector are labelled with color 'labelCol2'.}

\item{labelCol1}{color of genes in geneList1.}

\item{labelCol2}{color of genes in geneList2.}

\item{cond1Title}{plot title for the first condition.}

\item{cond2Title}{plot title for the second condition.}

\item{...}{other parameters to be passed to \code{\link{qgraph}}.}
}
\value{
a list of two "qgraph" objects, each corresponds to one of the
correlation plots of the two input conditions. See \code{\link{qgraph}}
for more details.
}
\description{
Plot differential co-expression network for a given set of genes.
}
\details{
This function plots correlation networks for a given set of genes separately
for two conditions. Edge weights represent the correlation coefficient
between two genes. The users can set various thresholds and colors to focus
on key changes in the network and understand the change in associations
between gene pairs under different conditions.
}
\examples{

# get expression data
data(exprBC)
# get sample information
data(samplesBC)
# get example genes
data(genesER)

# calculate differential correlation matrix
dcMat <- DCMat(exprMat = exprBC,
               condition = samplesBC$er_status,
               corMethod = "spearman")

# plot differential co-expression network using example genes
p <- DCPlot(exprMat = exprBC,
            condition = samplesBC$er_status,
            genes = genesER,
            dcMat = dcMat,
            posDCCut = 5,
            negDCCut = -5,
            corMethod = "spearman",
            minimum = 0.2)

}
\seealso{
\code{\link{DCMat}}

\code{\link{eigenCentrality}}
}
\author{
Lanying Wei
}
