% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprBC}
\alias{exprBC}
\title{Expression data from breast cancer patients}
\format{
A matrix with 12,787 rows (genes) and 400 columns (samples).
}
\usage{
data(exprBC)
}
\description{
This matrix contains a subset of RNA-seq gene expression data from patients
with breast cancer. The full data is available in the Gene Expression
Omnibus (GEO) repository under accession number GSE96058. The format of the
expression value is log2(FPKM + 0.1) and was rounded to one digit to reduce
the data size. The corresponding sample information is stored in the
data object "samplesBC".
}
\keyword{datasets}
