% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genesER}
\alias{genesER}
\title{A subset of genes involved in early estrogen response}
\format{
A character vector
}
\usage{
data(genesER)
}
\description{
This vector contains a subset of genes involved in early estrogen response.
}
\keyword{datasets}
