% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCosi.R
\name{getCosi}
\alias{getCosi}
\title{Cosinet Score}
\usage{
getCosi(
  exprMat,
  condition,
  genes,
  sampleNames = colnames(exprMat),
  bin = 0.1,
  binRange = 1,
  dcMat = NULL,
  posDCCut = 3,
  negDCCut = -posDCCut,
  weighted = TRUE,
  threads = 2
)
}
\arguments{
\item{exprMat}{an expression matrix. Rows represent genes and
columns represent samples.}

\item{condition}{a vector that specifies the condition of each sample.
The vector should contain only two conditions. The order of the vector
should match the column order of the expression matrix.}

\item{genes}{a character vector of genes to be evaluated. These genes can be
the genes involved in a function-specific differential co-expression network.
Gene pairs generated from this gene list that pass the thresholds for
dcMat are used to calculate Cosinet scores.}

\item{sampleNames}{a character vector of sample names to calculate the
Cosinet scores for.}

\item{bin}{a numeric value between 0 and 1 that represents the fraction of
the bin size relative to the binRange length. Binned regions are used to
estimate gene pair dependency at a local region around an individual sample.}

\item{binRange}{a numeric value between 0 and 1 that represents the fraction
of the middle expression range relative to the full expression range. The
length of binRange is used as the base to generate the bin length. When
binRange is set to 1, the range between the maximum expression value and
the minimum expression value is used as the base. If users are concerned
about extreme data having a large impact on the bin size setting, a smaller
value of binRange, for example 0.95, can be set so that the expression range
between the 2.5\% quantile and the 97.5\% quantile is used as the base. A
binRange value less than 0.9 is not recommended, as this will exclude up to
10\% of the samples from the bin size calculation.}

\item{dcMat}{a matrix of z-scores representing differential associations
of gene expression between two conditions. See {\link{DCMat}} for more
details. This matrix, along with posDCCut and negDCCut, are used to 1. remove
gene pairs that show low level of differential co-expression, so that only
important changes in the network are considered in the Cosinet score
calculation; and 2. used as weights when calculating Cosinet scores.}

\item{posDCCut}{a positive numeric value that serves as positive
threshold for dcMat. Gene pairs with dcMat values
below posDCCut and above negDCCut are not used for the calculation.}

\item{negDCCut}{a negative numeric value that serves as negative
threshold for dcMat. Gene pairs with dcMat values
below posDCCut and above negDCCut are not used for the calculation.}

\item{weighted}{logical indicator for weighting each gene pair by the
absolute value of corresponding dcMat score during Cosinet score calculation.}

\item{threads}{number of threads used for calculation.}
}
\value{
a matrix of the final Cosinet scores and the individual scores for
each gene pair before averaging. A lower Cosinet score indicates that the
gene co-expression patterns of the given sample are more similar to those of
the first condition, while a higher Cosinet score indicates that they are
more similar to those of the second condition.
}
\description{
Calculate Cosinet scores with a given set of genes.
}
\details{
Calculate Cosinet scores with a given set of genes. Cosinet uses gene
expression data to determine the degree of similarity between the
gene co-expression patterns of a given sample and reference conditions
in the context of a function-specific differential co-expression network.
Cosinet scores quantifies the degree of rewiring of co-expression network at
the level of individual samples.
}
\examples{

# get expression data
data(exprBC)
# get sample information
data(samplesBC)
# get example genes
data(genesER)

# calculate differential correlation matrix
dcMat <- DCMat(exprMat = exprBC,
               condition = samplesBC$er_status,
               corMethod = "spearman")

# calculate Cosinet scores with example genes
scoreMat <- getCosi(exprMat = exprBC,
                    condition = samplesBC$er_status,
                    genes = genesER,
                    bin = 0.1,
                    dcMat = dcMat,
                    posDCCut = 3,
                    negDCCut = -3,
                    threads = 2)


}
\seealso{
\code{\link{DCMat}}

\code{\link{eigenCentrality}}
}
\author{
Lanying Wei
}
