% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPreprocess.R
\name{matPreprocess}
\alias{matPreprocess}
\title{Expression Matrix Preprocessing}
\usage{
matPreprocess(
  exprMat,
  condition,
  minSample = 20,
  minExp = 0,
  logTrans = FALSE,
  logBase = 2,
  pseudoCount = 1
)
}
\arguments{
\item{exprMat}{an expression matrix. Rows represent genes and
columns represent samples.}

\item{condition}{a vector that specifies the condition of each sample.
The vector should contain only two conditions. The order of the vector
should match the column order of the expression matrix.}

\item{minSample}{minimum number of samples required to have an expression
value over minExp for each condition. Genes that fail this criterion are
removed.}

\item{minExp}{minimum expression value for gene filtration.}

\item{logTrans}{logical indicator for performing log-transformation.}

\item{logBase}{the base with respect to which logarithms are computed.}

\item{pseudoCount}{pseudo count to add before log-transformation.}
}
\value{
a matrix of prepossessed gene expression values.
}
\description{
Preprocess expression matrix for co-expression analysis.
}
\details{
Preprocess expression matrix for co-expression analysis. Perform
log-transformation, remove rows with duplicate gene names
and filter genes expressed in too few samples.
}
\examples{

testMat <- matrix(data = 0:11, nrow = 3)
matPreprocess(exprMat = testMat,
              condition = c(1, 1, 2, 2),
              minSample = 2, minExp = 1, logTrans = TRUE,
              logBase = 2, pseudoCount = 0.1)

}
\author{
Lanying Wei
}
