% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPlot.R
\name{rankPlot}
\alias{rankPlot}
\title{Rank Plot}
\usage{
rankPlot(
  geneScore,
  genesToLabel = NULL,
  nLabel = 10,
  nIQR = 1.5,
  labelSize = 3,
  main = NULL
)
}
\arguments{
\item{geneScore}{a numeric vector of scores to rank, with gene as names.}

\item{genesToLabel}{a character vector specifying the genes to be labeled.}

\item{nLabel}{an integer specifying the number of top scored genes to be
labeled.}

\item{nIQR}{a positive numeric value indicating the fold of IQR to be used as
cutoff to color genes. Genes with scores > Q3 + nIQR * (Q3 - Q1) are colored
red.}

\item{labelSize}{a positive numeric value specifying the size of the label.}

\item{main}{a character string representing the plot title.}
}
\value{
A \code{ggplot} scatter plot object showing the ranked gene scores.
}
\description{
Draw the score and rank of genes on a scatter plot.
}
\examples{
# get expression data
data(exprBC)
# get sample information
data(samplesBC)

# calculate differential correlation matrix between ER status
dcMat <- DCMat(exprMat = exprBC,
               condition = samplesBC$er_status,
               corMethod = "spearman")

# calculate gene centrality base on differential correlation matrix
geneCentrality <- eigenCentrality(mat = dcMat)

# plot gene centrality rank plot
rankPlot(geneCentrality)

}
\seealso{
\code{\link{eigenCentrality}}
}
\author{
Lanying Wei
}
