% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplesBC}
\alias{samplesBC}
\title{Phenotype data from breast cancer patients.}
\format{
A data frame containing 11 variables:
\describe{
\item{title}{Sample title.}
\item{age_at_diagnosis}{Age at diagnosis.}
\item{chemo_treated}{Whether the patient received adjuvant chemotherapy
(1 for "yes", 0 for "no").}
\item{endocrine_treated}{Whether the patient received adjuvant endocrine
therapy (1 for "yes", 0 for "no").}
\item{er_status}{Estrogen receptor status (1 for "positive", 0 for
"negative").}
\item{her2_status}{Human epidermal growth factor receptor 2 (HER2) status
(1 for "positive", 0 for "negative").}
\item{pgr_status}{Progesterone receptor status (1 for "positive", 0 for
"negative").}
\item{nhg}{Nottingham histological grade.}
\item{pam50_subtype}{PAM50 subtype.}
\item{overall_survival_days}{Days of overall survival.}
\item{overall_survival_event}{Overall survival event (1 for "deceased", 0
for "alive").}
}
}
\usage{
data(samplesBC)
}
\description{
This data frame contains a subset of phenotype data from patients with
breast cancer. Full data are available in the Gene Expression Omnibus
(GEO) repository under accession number GSE96058. The corresponding
expression data is stored in the data object "exprBC".
}
\keyword{datasets}
